Fix non-determinism in sentence imposed group summary (Recidiviz/recidiviz-data#16404)
## Description of the change

While validating output from `end_date_exclusive` changes, I found
another source of non-determinism. In our sentence group summary view we
create booleans that are used to identify the most severe sentence and
the first sentence within a group. However, in states that have more
than 1 charge associated with the same sentence, our current output is
non-deterministic in cases where more than 1 charge associated with the
same sentence are of equal sentence length (tie for most severe).

This PR adds `charge_id` to the order by so that we deterministically
choose the most severe in these cases.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: df516c5257c1788718d2f33044b864f40b993b27