[Justice Counts][2/n] Justice Counts-specific `deploy_to_staging.sh` script (Recidiviz/recidiviz-data#16144)
## Description of the change

After much ado, here is a draft of a deploy script for Justice Counts.
The script performs the following actions:

1) Runs our Cloud Build Trigger to build a Docker image off of main in
`recidiviz-data` and `justice-counts` (See [previous
PR](https://github.com/Recidiviz/recidiviz-data/pull/16370) in the stack
for the script that does this.)
2) Runs migrations on `recidiviz-staging` to head of main. (This PR also
takes the Justice Counts database migrations out of the regular data
platform deploy).
3) Deploys a new Cloud Run revision and allocates 100% traffic.
4) Tags the commits (both backend and frontend) that were used in the
deploy.

There are still several TODOs here, some of which are marked inline, but
I wanted to get the skeleton functionality out first to see if I was on
the right track with this.

After this, I will create a `deploy_to_production.sh` script that copies
the Docker image created in step 1 to `recidiviz-123`, runs steps 2-3 on
`recidiviz-123`, and then creates a new release in Github.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Contributes to Recidiviz/recidiviz-data#16325

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

Co-authored-by: Helper Bot <helperbot@recidiviz.org>
GitOrigin-RevId: 5a4e9ed1397cd445671d4ab2ae762aa50a4e867e