[Workflows] filter CR records for remaining_criteria_needed (Recidiviz/recidiviz-data#16405)
## Description of the change

Filter Compliant Reporting referral records such that only records here
`remaining_criteria_needed <= 1`, which is the criteria for almost
eligible is true. This ensures that we don't have opportunity records in
the frontend that must be filtered out, which has turned out to be bug
prone.

This has been loaded into a sandbox - jovergaag_2558_ - there are now
2764 records instead of 4798. Does that seem like the correct number of
eligible/almost eligible records given this new filter?

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes
[recidiviz-dashboards/2558](https://github.com/Recidiviz/recidiviz-dashboards/issues/2558)

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f1a1e7cdfc075a911d7ba988e234205631f26883