[US_MI] Ingest state assessments (Recidiviz/recidiviz-data#16323)
## Description of the change

Ingest state assessments for MI.

MI categorizes the assessments they conduct into two categories: COMPAS
assessments (which are developed and owned by Northpointe/Equivant) and
non-COMPAS assessments (such as STATIC and STABLE). COMPAS has its own
database (the COMPAS database), so all COMPAS assessments data is stored
there and will continue to be entered/stored there even with the
introduction of COMS (the new data system that's coming). Some
non-COMPAS assessments are also stored in the COMPAS database, but some
non-COMPAS assessments are stored in other databases as well, and
there's no standardized policy for where they should be stored.
Furthermore, with the introduction of COMS, some non-COMPAS assessments
data is getting stored there too. This query gathers all the assessment
information that can be found in the COMPAS database, which should be
sufficient for our current needs since supervision level decisions (for
Workflows) only currently depend on COMPAS assessments (specifically the
VFO recidivism risk and non-VFO recidivism risk scales). But if we need
other assessments, we might need to request a lot more data and refactor
this.

STILL TO DO:

- [x] When StateAssessmentType.COMPAS gets implemented, map COMPAS
assessments to StateAssessmentType.COMPAS instead of
StateAssessmentType.INTERNAL_UNKNOWN

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 6ed2316c1ba52373c4227e2df944dd55c6a6f36c