[US_ID][Workflows] Exclude clients at liberty and use most recent FTRD instead of null (Recidiviz/recidiviz-data#16482)
## Description of the change

Two bugs were found:
1. Clients were showing up in TES eligibility queries who were not on
supervision, because a new supervision period is created due to a new PO
assignment.
2. Some clients have COMPLETED sentences but still appear to be on
supervision. In this case, their FTRD is NULL due to no open sentence
span, so they are showing up as almost eligible for LSU and ED when they
should be eligible for overdue for discharge.

1. To address the first bug, I created the `NOT_AT_LIBERTY` task, which
changed eligibility numbers below:

no one that is currently eligible for LSU/ED is not on supervision, but
~200 individuals that were showing up as overdue for discharge had
already been discharged.
- LSU: no change (543)
- ED: no change (778)
- FT: 466 --> 273

2. For the second bug, we set the most recent
`projected_completion_date_max_spans` span end_date to NULL, so that the
most recent full term release date will be used and those clients will
show up as eligible as overdue for discharge.

There were 3 cases where a client appeared as eligible for LSU/ED where
they should have been eligible for overdue for discharge, but the
majority of cases were new clients that are currently on supervision but
for some reason have completed sentence spans.
- LSU: 542
- ED: 777
- FT: 320
There are 5 clients who have no supervising_officer_external_id

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0666efd281c245294cdef48fe890181fa3730650