Sessions `end_date_exclusive` field (Recidiviz/recidiviz-data#16352)
## Description of the change

This PR adds a field `end_date_exclusive` to all of our views in the
sessions dataset that represent spans. All views (with the exception of
the 5 population preprocessing files) continue to have a field called
`end_date` as well. In most sessions views, they are currently end date
exclusive, meaning that the `end_date` will be 1 day less than the
`end_date_exclusive`. However, some of our sessions views are already
end date exclusive in which case the two end date fields will be the
same.

Changes to use `end_date_exclusive` were made upstream in each view's
query and then `end_date` is just added at the end. This includes all
logic to calculate session length, session overlaps, joins with events,
etc. This should make deprecating `end_date` much easier because it is
already removed from all of logic within the sessions dataset.

**Validation**
This change should have no impact on the overall sessions output. I
tested that in
[this](https://github.com/Recidiviz/recidiviz-research/blob/main/agaidus/2022-11-03%20-%20Compare%20End%20Date%20Exclusive%20Refactor%20Output.ipynb)
notebook. I wrote a function that takes a sandbox prefix, dataset, and
table name and compares the version in the sandbox to the version in
staging. The function looks for exact value matches between every row
and column (for column types that are dates, integers, or strings)/
Column names are pulled from the staging version of the view so the new
`end_date_exclusive` field is not factored into this comparison.

Every view in the sessions dataset is compared, which gives confidence
that any downstream uses of sessions will not be impacted by this
change.

The function compares two sandbox datasets `agaidus7` (version off of
this branch) vs `agaidus0` (version off of main). I ran a sandbox off of
main because I made recent changes in a separate PR that were not yet
reflected in staging.

Out of the 50 views that were checked, the following 3 are the ones that
do not have a perfect 100% match. However, all match rates are greater
than 99.99% so I am comfortable moving forward with this. I suspect that
the remaining differences are due to non-determinism in certain fields
that is unrelated to this PR.

<img width="598" alt="Screen Shot 2022-11-03 at 1 48 28 PM"
src="https://user-images.githubusercontent.com/6909506/199830942-b6e64fea-9171-42e9-8177-3244395dbcea.png">

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#14675

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8bfaf5943c221173dc416ca6bd8c2b0afec90ad9