[Justice Counts] Justice Counts-specific `deploy_to_prod` script (Recidiviz/recidiviz-data#16443)
## Description of the change

This PR is a companion to
https://github.com/Recidiviz/recidiviz-data/pull/16144, which deploys to
staging. After we've used that script to deploy to staging, we can
deploy to prod as follows:

1) Moves the Docker image wwith the specified version tags in
recidiviz-staging to recidiviz-123.
2) Runs migrations on recidiviz-123 to the commit that this Docker image
was built on. (We can determine this commit by looking at the first tag
on this image.)
3) Deploys a new Cloud Run revision and allocates 100% traffic.
(Straightforward!)

Just tested this out and it seems good! LMK if I'm missing anything. I
do want to automatically create a release in the justice-counts repo at
the end of this script, but will do that in a followup.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Contributes to Recidiviz/recidiviz-data#16325

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

Co-authored-by: Helper Bot <helperbot@recidiviz.org>
GitOrigin-RevId: bce35320c181a08558678b444e647794aeb86428